Move multiple graph acceptance tests into behaviour
